backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
transformers==4.37.2
torch==2.2.0
torchaudio==2.2.0
numpy<2
